Combined form-types GET endpoint controller with Case Status main endpoint controller
