+++ SocialCareCaseViewerApi/Startup.cs
@@ -152,7 +152,6 @@ namespace SocialCareCaseViewerApi
             services.AddScoped<IRelationshipsUseCase, RelationshipsUseCase>();
             services.AddScoped<IPersonalRelationshipsUseCase, PersonalRelationshipsUseCase>();
             services.AddScoped<ICaseStatusesUseCase, CaseStatusesUseCase>();
-            services.AddScoped<IGetCaseStatusFieldsUseCase, GetCaseStatusFieldsUseCase>();
             services.AddScoped<ICreateRequestAuditUseCase, CreateRequestAuditUseCase>();
         }
