--- SocialCareCaseViewerApi/V1/UseCase/GetCaseStatusFieldsUseCase.cs
@@ -1,36 +1,0 @@
-using SocialCareCaseViewerApi.V1.Boundary.Requests;
-using SocialCareCaseViewerApi.V1.Boundary.Response;
-using SocialCareCaseViewerApi.V1.Exceptions;
-using SocialCareCaseViewerApi.V1.Factories;
-using SocialCareCaseViewerApi.V1.Gateways;
-using SocialCareCaseViewerApi.V1.UseCase.Interfaces;
-
-namespace SocialCareCaseViewerApi.V1.UseCase
-{
-    public class GetCaseStatusFieldsUseCase : IGetCaseStatusFieldsUseCase
-    {
-        private readonly IDatabaseGateway _databaseGateway;
-
-        public GetCaseStatusFieldsUseCase(IDatabaseGateway databaseGateway)
-        {
-            _databaseGateway = databaseGateway;
-        }
-
-        public GetCaseStatusFieldsResponse Execute(GetCaseStatusFieldsRequest request)
-        {
-            var caseStatusType = _databaseGateway.GetCaseStatusTypeWithFields(request.Type);
-
-            if (caseStatusType == null)
-            {
-                throw new CaseStatusNotFoundException();
-            }
-
-            return new GetCaseStatusFieldsResponse
-            {
-                Description = caseStatusType.Description,
-                Name = caseStatusType.Name,
-                Fields = caseStatusType?.Fields.ToResponse()
-            };
-        }
-    }
-}
