+++ SocialCareCaseViewerApi/V1/UseCase/Interfaces/ICaseStatusesUseCase.cs
@@ -6,6 +6,8 @@ namespace SocialCareCaseViewerApi.V1.UseCase.Interfaces
 {
     public interface ICaseStatusesUseCase
     {
+        GetCaseStatusFieldsResponse ExecuteGetFields(GetCaseStatusFieldsRequest request);
+
         ListCaseStatusesResponse ExecuteGet(long personId);
         CaseStatus ExecutePost(CreateCaseStatusRequest request);
     }
